Fix: Card 눌렀을 때 URL 변경 및 util 함수 파일 변경
